pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vtes-twd-scraper"
version = "0.1.0"
description = "Scrape VTES tournament winning decks from vekn.net and export to YAML"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }

dependencies = [
    "httpx>=0.27",           # HTTP client (async-ready for future extension)
    "beautifulsoup4>=4.12",  # HTML parsing
    "lxml>=5.0",             # Fast HTML parser backend for bs4
    "pydantic>=2.7",         # Data validation and models
    "ruamel.yaml>=0.18",     # YAML serialization (preserves formatting)
    "typer>=0.12",           # CLI framework
    "rich>=13.0",            # Pretty terminal output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",             # Linter + formatter
    "mypy>=1.10",
]

[project.scripts]
vtes-scraper = "vtes_scraper.cli:main"

[tool.setuptools.packages.find]
include = ["vtes_scraper*"]

# ---------------------------------------------------------------------------
# Ruff (linter/formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
